+++ CHANGELOG.md
@@ -3,8 +3,12 @@
 ### Fixes
 - `from_regex/1`:
     - Ignore inline comments `(?# )`
+
+### New features
+- `from_regex/1`:
     - Add support for `:extended` (`/x`) modifier
     - Add support for `:multiline` (`/m`) modifier
+    - Add support for `:firstline` (`/f`) modifier
 
 ## 0.2.0 [2026-03-30]